requirements.txt
mkdocs==1.5.3
mkdocs-material==9.5.15
mkdocs-git-revision-date-localized-plugin==1.2.4
mkdocs-git-authors-plugin==0.8.0
nltk==3.8.1
# mkdocs-pdf-export-plugin
git+https://github.com/yuk1i/mkdocs-with-pdf.git
GitPython